Initial rock climbing page
Create initial rock climbing page, that includes a feature
video, and a sidebar with links to additional send videos.

The page is not accessible from the main page yet until a
navigation bar is created, at which point it will be
accessible and the navigation bar can be added to the
rock climbing page and the about-me page.