Modify about me and climbing pages
Update the about me and change the color of the Baltimore link
to purple instead of orange.

Change the layout of the climbing page to have the links to
send videos be below the featured video.

Change the featured climbing video the Occam's Razor.

Add links to more climbing videos.

Create gitignore file