Add explore and music page, reorg climbing page
Create the Explore page that will be part of the main navigation,
and will be the page that contains the misc. stuff I want share,
so that the main navigation bar stays small and simple.
- Use photo/title as link to go to the topic pages.

Create the music page, that currently just lists my favorite
songs of the past years using embedded youtube videos of the
songs.

Reorganize the climbing page to just be embedded youtube videos
of all my recorded sends, in reverse chronological order.